+++ webapp/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.4.9
+
+- Improve
+  - HTTPレスポンス時、明示的に Content-Type を指定するよう修正
+
 ## 1.4.8
 
 - Improve